refactor: rename NoIconsFound component file and fix type error in iconCard
